requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.0
scipy==1.15.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
gunicorn==20.1.0
